Move helper functions to test utils
